requirements.txt
numpy
easyocr
sklearn
flask
scikit-learn
Pillow==9.5.0
selenium
tqdm
pandas
autolog